+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * `Versions`:
   * `Exposed`: `0.59.0` -> `0.60.0`
+* `Repo`:
+  * `Cache`:
+    * Add extensions `alsoInvalidate` and `alsoDoInvalidate`
+* `Koin`:
+  * Add extensions `singleSuspend` and `factorySuspend` for defining of dependencies with suspendable blocks
 
 ## 0.25.1